Split `api/v1/views.py`, it's getting long.
Fixes #111

We now have:
  - base (generics, mixins, abstracts)
  - iam (user/group/namespace/permissions)
  - attachments
  - extensions
  - resources
