+++ hubuum/api/v1/views/__init__.py
@@ -1,0 +1,1 @@
+"""The hubuums.api.v1.views package."""
