+++ hubuum/api/v1/views/resources.py
@@ -1,0 +1,137 @@
+"""Views for the resources in API v1."""
+
+from hubuum.api.v1.serializers import (
+    HostSerializer,
+    HostTypeSerializer,
+    JackSerializer,
+    PersonSerializer,
+    PurchaseOrderSerializer,
+    RoomSerializer,
+    VendorSerializer,
+)
+from hubuum.filters import (
+    HostFilterSet,
+    HostTypeFilterSet,
+    JackFilterSet,
+    PersonFilterSet,
+    PurchaseOrderFilterSet,
+    RoomFilterSet,
+    VendorFilterSet,
+)
+from hubuum.models.resources import (
+    Host,
+    HostType,
+    Jack,
+    Person,
+    PurchaseOrder,
+    Room,
+    Vendor,
+)
+
+from .base import HubuumDetail, HubuumList
+
+
+class HostList(HubuumList):
+    """Get: List hosts. Post: Add host."""
+
+    queryset = Host.objects.all().order_by("id")
+    serializer_class = HostSerializer
+    filterset_class = HostFilterSet
+
+
+class HostDetail(HubuumDetail):
+    """Get, Patch, or Destroy a host."""
+
+    queryset = Host.objects.all()
+    serializer_class = HostSerializer
+    lookup_fields = ("id", "name", "fqdn")
+
+
+class HostTypeList(HubuumList):
+    """Get: List hosttypes. Post: Add hosttype."""
+
+    queryset = HostType.objects.all().order_by("name")
+    serializer_class = HostTypeSerializer
+    filterset_class = HostTypeFilterSet
+
+
+class HostTypeDetail(HubuumDetail):
+    """Get, Patch, or Destroy a hosttype."""
+
+    queryset = HostType.objects.all()
+    serializer_class = HostTypeSerializer
+
+
+class RoomList(HubuumList):
+    """Get: List rooms. Post: Add room."""
+
+    queryset = Room.objects.all().order_by("id")
+    serializer_class = RoomSerializer
+    filterset_class = RoomFilterSet
+
+
+class RoomDetail(HubuumDetail):
+    """Get, Patch, or Destroy a room."""
+
+    queryset = Room.objects.all()
+    serializer_class = RoomSerializer
+
+
+class JackList(HubuumList):
+    """Get: List jacks. Post: Add jack."""
+
+    queryset = Jack.objects.all().order_by("name")
+    serializer_class = JackSerializer
+    filterset_class = JackFilterSet
+
+
+class JackDetail(HubuumDetail):
+    """Get, Patch, or Destroy a jack."""
+
+    queryset = Jack.objects.all()
+    serializer_class = JackSerializer
+
+
+class PersonList(HubuumList):
+    """Get: List persons. Post: Add person."""
+
+    queryset = Person.objects.all().order_by("id")
+    serializer_class = PersonSerializer
+    filterset_class = PersonFilterSet
+
+
+class PersonDetail(HubuumDetail):
+    """Get, Patch, or Destroy a person."""
+
+    queryset = Person.objects.all()
+    serializer_class = PersonSerializer
+
+
+class VendorList(HubuumList):
+    """Get: List vendors. Post: Add vendor."""
+
+    queryset = Vendor.objects.all().order_by("vendor_name")
+    serializer_class = VendorSerializer
+    filterset_class = VendorFilterSet
+
+
+class VendorDetail(HubuumDetail):
+    """Get, Patch, or Destroy a vendor."""
+
+    queryset = Vendor.objects.all()
+    serializer_class = VendorSerializer
+
+
+class PurchaseOrderList(HubuumList):
+    """Get: List purchaseorders. Post: Add purchaseorder."""
+
+    queryset = PurchaseOrder.objects.all().order_by("id")
+    serializer_class = PurchaseOrderSerializer
+    filterset_class = PurchaseOrderFilterSet
+
+
+class PurchaseOrderDetail(HubuumDetail):
+    """Get, Patch, or Destroy a purchaseorder."""
+
+    queryset = PurchaseOrder.objects.all()
+    serializer_class = PurchaseOrderSerializer
